+++ extension/src/content-script/osu/pages/friends.ts
@@ -1,5 +1,7 @@
-import { addFlagUser } from "@src/content-script/osu/flagHtml";
-import { idFromProfileUrl, nextAbortControllerSignal } from "./content";
+import { addFlagUsers } from "@src/content-script/osu/flagHtml";
+import { TFlagItems } from "@src/utils/html";
+import { idFromOsuProfileUrl } from "@src/utils/utils";
+import { nextAbortControllerSignal } from "@src/utils/fetchUtils";
 
 // https://osu.ppy.sh/home/friends
 const setActualFriendsObserver = new MutationObserver(() => {
@@ -39,12 +41,13 @@ export const updateFlagsFriends = async () => {
 
     const friendsList = document.querySelector(".user-list")!.querySelectorAll(".user-card__details");
 
+    const flagItems: TFlagItems = [];
     for (const item of friendsList) {
-        if (signal.aborted) {
-            return;
-        }
+
         const playerNameElement = item.querySelector(".user-card__username") as HTMLElement;
-        const playerId = idFromProfileUrl(playerNameElement.getAttribute("href")!);
-        await addFlagUser(item as HTMLElement, playerId, { addDiv: true, addMargin: true, signal: signal });
+        const playerId = idFromOsuProfileUrl(playerNameElement.getAttribute("href")!);
+        flagItems.push({ id: playerId, item: item as HTMLElement });
     }
+    await addFlagUsers(flagItems, { addDiv: true, addMargin: true, signal: signal });
+
 };