+++ extension/src/content-script/osu/pages/profile.ts
@@ -2,7 +2,8 @@ import { addFlagUser } from "@src/content-script/osu/flagHtml";
 // https://osu.ppy.sh/users/4871211/fruits
 
 import { isNumber } from "@src/utils/utils";
-import { idFromProfileUrl, nextAbortControllerSignal } from "./content";
+import { idFromOsuProfileUrl } from "@src/utils/utils";
+import { nextAbortControllerSignal } from "@src/utils/fetchUtils";
 import { osuScoreRanking } from "@src/utils/respektive";
 import {
     getActiveLanguageCode,
@@ -10,6 +11,8 @@ import {
     getLocMsg,
     waitLastLanguageIsLoaded,
 } from "@src/utils/languagesChrome";
+import osuNameToCode from "../osuNameToCode";
+import { getCountryName } from "@src/utils/flagsJsonUtils";
 
 export const profileMutationObserverInit = new MutationObserver((_) => {
     updateFlagsProfile();
@@ -32,7 +35,7 @@ export const updateFlagsProfile = async () => {
     });
 
     const url = location.href;
-    const playerId = idFromProfileUrl(url);
+    const playerId = idFromOsuProfileUrl(url);
     if (!isNumber(playerId)) {
         return;
     }
@@ -44,11 +47,20 @@ export const updateFlagsProfile = async () => {
     addScoreRank(signal);
     const flagResult = await addFlagUser(flagElement as HTMLElement, playerId, { signal: signal, addMargin: true });
     if (!flagResult) return;
-    const { countryName, regionName } = flagResult;
+    const { countryCode, countryName, regionName } = flagResult;
     const countryNameElement = flagElement.querySelector(".profile-info__flag-text")!;
+    
+    let countryText = countryNameElement.textContent?.split(" |")[0];
+    const originalCountryCode = osuNameToCode(countryText!);
 
-    const originalCountry = countryNameElement.textContent?.split(" / ")[0];
-    const replaceText = `${countryName ? countryName : originalCountry}${regionName ? ` / ${regionName}` : ""}`;
+    let replaceText: string;
+    if (originalCountryCode === countryCode){
+        countryText = countryNameElement.textContent?.split(" / ")[0];
+        replaceText = `${countryName ? countryName : countryText}${regionName ? ` / ${regionName}` : ""}`;
+    }else{
+        const regionCountryName = await getCountryName(countryCode!);
+        replaceText = `${countryName} | ${regionCountryName} / ${regionName}`;
+    }
 
     countryNameElement.textContent = replaceText;
 };